refactor(#792): split local_dispatcher.rs into single-responsibility files
local_dispatcher.rs was 633 lines carrying six responsibilities. Extracted each
into its own file (file-per-feature, matching the existing local_dispatcher_<x>
handlers), leaving the core (struct + new + dispatch router) at 400 lines:
- local_dispatcher_attach.rs    — dependency-attach setters (wiring)
- local_dispatcher_queries.rs   — read accessors / state exposure
- local_dispatcher_access.rs    — with_block / with_chain borrow helpers
- local_dispatcher_midi_system.rs — handle_midi_system (+ its tests)
- local_dispatcher_paths.rs     — handle_paths_system

Pure move — byte-identical behaviour, same public API. The 3 handle_midi_system
tests moved to local_dispatcher_midi_system_tests.rs (each file with its test).
Fixed the stale 'Phase 1 skeleton unimplemented!' module doc (audit #14).
local_dispatcher.rs leaves the debt list. 237/237 application tests pass; static
gate PASSED. Unblocks the config-bus fixes (#8-#16), which need a new dispatcher
field the cap previously refused.
