+++ crates/application/src/lib.rs
@@ -18,6 +18,8 @@ pub mod di_loader;
 pub mod dispatcher;
 pub mod event;
 pub mod local_dispatcher;
+mod local_dispatcher_access;
+mod local_dispatcher_attach;
 mod local_dispatcher_block_edit;
 mod local_dispatcher_block_lifecycle;
 mod local_dispatcher_block_param;
@@ -31,7 +33,10 @@ mod local_dispatcher_diagnostic;
 mod local_dispatcher_io_binding;
 mod local_dispatcher_ir_reseed;
 mod local_dispatcher_language;
+mod local_dispatcher_midi_system;
 mod local_dispatcher_output;
+mod local_dispatcher_paths;
+mod local_dispatcher_queries;
 mod local_dispatcher_plugin_catalog;
 mod local_dispatcher_preset;
 mod local_dispatcher_project;
