+++ crates/application/src/local_dispatcher_tests.rs
@@ -3369,34 +3369,6 @@ pub(super) fn empty_project_rc() -> Rc<RefCell<Project>> {
 }
 
 #[test]
-fn save_midi_devices_emits_event_without_mutating_project() {
-    let project = empty_project_rc();
-    let before = project.borrow().clone();
-    let dispatcher = LocalDispatcher::new(Rc::clone(&project));
-
-    let events = dispatcher
-        .dispatch(Command::SaveMidiDevices { devices: vec![] })
-        .unwrap();
-
-    assert_eq!(events, vec![Event::MidiDevicesSaved]);
-    assert_eq!(
-        project.borrow().chains.len(),
-        before.chains.len(),
-        "system command must not touch project chains"
-    );
-    assert_eq!(
-        project.borrow().device_settings.len(),
-        before.device_settings.len(),
-        "system command must not touch project device_settings"
-    );
-    assert_eq!(
-        project.borrow().name,
-        before.name,
-        "system command must not touch project name"
-    );
-}
-
-#[test]
 fn save_midi_mapping_writes_bindings_into_project_midi() {
     let project = empty_project_rc();
     let dispatcher = LocalDispatcher::new(Rc::clone(&project));
@@ -3424,36 +3396,8 @@ fn save_midi_mapping_writes_bindings_into_project_midi() {
     assert_eq!(stored, bindings);
 }
 
-#[test]
-fn start_and_stop_midi_learn_emit_events() {
-    let project = empty_project_rc();
-    let dispatcher = LocalDispatcher::new(Rc::clone(&project));
-    assert_eq!(
-        dispatcher.dispatch(Command::StartMidiLearn).unwrap(),
-        vec![Event::MidiLearnStarted]
-    );
-    assert_eq!(
-        dispatcher.dispatch(Command::StopMidiLearn).unwrap(),
-        vec![Event::MidiLearnStopped]
-    );
-}
-
-#[test]
-fn publish_midi_event_passthrough_emits_midi_event_received() {
-    let project = empty_project_rc();
-    let dispatcher = LocalDispatcher::new(Rc::clone(&project));
-    let source = project::midi::Source::Cc {
-        channel: 1,
-        controller: 7,
-    };
-    let events = dispatcher
-        .dispatch(Command::PublishMidiEvent {
-            source: source.clone(),
-        })
-        .unwrap();
-    assert_eq!(events, vec![Event::MidiEventReceived { source }]);
-}
-
+// #513 / #493: System / MIDI tests moved to
+// `local_dispatcher_midi_system_tests.rs` (each file with its test).
 // #513 / #540: System / Paths (presets + plugins) tests moved to
 // `local_dispatcher_paths_tests.rs` so this file does not grow further
 // (already over the per-file size cap) and so the FS-sandboxing helper
